feat(bootstrap): always check for root and inbox dir
This commit adds functionality to always check for the presence of the
root and inbox directories and attempts to create them should they not
exist.

I'm aware that error handling here may not be idiomatic but this is
something I'll look to improve once I've got a minimal working product.