feat(new): add `new` command to create new notes
This creates notes using my existing format with a slight modification.

All notes contain a unique id (timestamp based), in my
current setup I only include this in the frontmatter of a note but not
in the filename. I've decided to include the timestamp as the prefix of
the filenames to make it easier to discern when a note was created
without having to open it.

I've also added the `-p` flag to mark notes as private i.e. end with
`.private.md` instead of `.md` this is so that I can have all notes in
one directory instead of two seperate ones based on visibility (public
and private). My personal use of this is to prevent publishing my
private notes anywhere that is publicly visible.