feat(link): add command to create wiki link to a note
This commit will add a command that generates a wiki link to an
interactively selected note. The output doesn't contain a trailing new
line leading to undesirable output in the terminal (the prompt is placed
directly after the output). Though, this is by design since it works
with my current vim setup.

If the behaviour causes any annoyance, I will look to add a flag to the
command that when present will add a trailing new line to the output. As
of now this isn't necessary since I don't anticipate using this command
outside of vim.