chore(ci): add release actions and remove faulty dry_run
Add an action to build and release `sb` for Linux and MacOS for both the
arm64 and amd64 architectures since these are the 4 that are relevant to
me.

The dry run option added previous was added to the wrong workflow in the
action, however, from what is saw in the commit and from the
documentation of the workflows I'm happy that it meets my needs.