fix(ci): add `token` key in the right place
I added an extra `with` under an existing one, this removes the extra
`with` and places the `token` key in the correct place.

This is why you don't do stuff when it's nearly midnight. So many extra
commits which I wouldn't have had to made if I wasn't trying to complete
this before I sleep.