feat: create files with kebab cased title
**What?**

This commit converts all titles into kebab case for their filenames -
the title within the body of the note remains unchanged.

**Why?**

The title could include any character present in the title of the note,
for example colons `:`. If the title contains characters reserved by the
filesystem this could be problematic. By using a neutral character set for
filenames such as alphanumeric characters and dashes, this utility
becomes more accessible to systems with various filesystems.