feat: create daily note before appending new note link
**What?**

If missing, create the daily note before appending a backlink to a new
note to it.

**Why?**

It's possible that a new (non-daily) note is created before the daily
note, in this scenario the backlink to the new note is appended to an
empty file. This means that the daily note is created without it's
template.

By making this change, the two code paths that lead to the daily note
being created will do so with it's template.