chore: Move `titleToKebabCase` to `internal`
Instead of keeping shared logic under `cmd/utils.go`, let's start moving
it to `internal`.

This means that the functions won't be publicly availabe if, for
whatever reason, someone decides to include this repository as a
dependency of their project.

As part of the move to `internal/` I have added tests, the plan is to do
this for all of the functions in `cmd/utils.go`.