chore(pre-commit): Add pre-commit config
This is a pretty simple configuration. The golang hooks, do provide a
testing hook, though due to the current lack of tests I have opted to
omit them for now.

I am also aware that the version of
<https://github.com/TekWizely/pre-commit-golang> is ~4 years old,
however, the alternative was
<https://github.com/dnephin/pre-commit-golang> which was last committed
to ~3 years ago and is now archived, whilst a v1.0.0 is being worked on
for the current option.

I was initially planning on include a linting hook, however, the
`golangci-lint` hooks provided don't seem to work for me. For now I'll
continue running this manually.