chore: Move `openFileInVim` to `internal`
Unlike the other functions, I have not added tests here.

This is due to my lack of knowledge around mocking in Go, which would
allow me to appropriately test code which runs shell commands.

I don't intend to leave this code in such a state and there are three
options that I can think of:

1. Keep the code as is, and write tests for it
  a. This is a no-go as it is difficult to test
2. Refactor the code, and write tests for it
  a. The best option if I want to keep this functionality, requires me
     to learn more about mocking in Go
3. Remove this code, and rely on editor integrations to interface with
   `sb`
   a. A consideration I've had, though it will lead to a worsened UX
      since it introduces friction to creating a new note when not
      already using the editor.
   b. On the other hand, it forces me to continue working on `sb` in a
      way that is not tied to a specific editor (currently building with
      mainly Vim and Neovim in mind)