feat: Add link command (#22)
* feat: Add `link` command

The `sb link` command allows the user to pass in a source (path) and destination
note (filename), displaying a relative markdown link unless `-w / --wiki` is
specified in which case a Wiki Link is displayed.

The intention of this command is to be used by editor integrations to
insert a link in the current note.

* chore: Add GitHub Action to run tests on PR

This should help prevent non-working code, or code which doesn't behave
as expected from making it into main.