feat: Updated daily command
This is part of a change of approach to the daily note. Rather than a
new note for each day, the idea is to have a single `daily.md`, or
grouped daily notes (e.g. per month or per year).

Instead of generating a note, the command now outputs an entry for a
daily note (in Neovim [or Vim] you can create keybind for `G:r!sb
daily<CR>` to add an entry to the end of the file)