feat: Improved linking behaviour
The main change made here, is to URL encode filenames which contains
spaces, making the navigation of these notes work as expected when viewing
them in GitHub.

The path command has been updated to URL decode the filenames when the
input is from a markdown link, a subsequent commit will add
functionality for finding the path of a note from a wikilink. It's key
to note that the desired functionality is that the uniqueness of
filenames will be case insensitive (to account for filesystems which
behave like this - looking at you APFS).

Additionally, the new command will be updated to allow spaces in
filenames, as well as some other functionality changes to reduce some of
the friction when creating notes across devices.