feat: Revised `new` command
The aim of this change is to reduce friction when creating notes, but
also offers aesthetic improvements when looking at graph views in
Obsidian.

Additionially, it removes the requirement of creating titles which are
kebab-cased when creating notes when not interfacing with the `sb` cli.

The new filename opinions being introduced by this change are:

  - Filenames may contain alphanumeric characters, spaces, hyphens, and
    underscores
  - Filenames must start and end with an alphanumeric character
    (excluding the `.md` extension)
  - Filenames should not contain consecutive spaces, hyphens, or
    underscores
  - Filenames should be unique regardless of casing
    * The existence of `Hello World.md` will prevent the creation of
      `hello world.md`