feat: `--date` flag for new command
This allows the creation of notes which include the current date in the
filename. Whilst, the Zettelkasten method includes providing unique
identifiers (usually the current date and time), this isn't as necessary
in a digital system where notes can be easily searched. However, there
is utility in dating notes where the creation time is key to identifying
them.